Upstream pull

+++ pengwin-setup.d/docker.sh
@@ -1,32 +1,50 @@
 #!/bin/bash
 
-source $(dirname "$0")/common.sh "$@"
+# shellcheck source=/usr/local/pengwin-setup.d/common.sh
+source "$(dirname "$0")/common.sh" "$@"
 
 DOCKER_VERSION="18.09.2"
 DOCKER_COMPOSE_VERSION="1.23.2"
+
+# Imported from common.sh
+declare wHome
+declare GOVERSION
+
+#Imported global variables
+declare USER
 
 function docker_install_build_relay() {
   #Build the relay
   if [[ ! -f "${wHome}/.npiperelay/npiperelay.exe" ]]; then
 
     echo "Checking for go"
-    if ! (go version); then
+    command_check '/usr/local/go/bin/go' 'version'
+    local go_check=$?
+    if [ $go_check -eq 1 ] ; then
       echo "Downloading Go using wget."
       wget -c "https://dl.google.com/go/go${GOVERSION}.linux-$(dpkg --print-architecture).tar.gz"
       tar -xzf go*.tar.gz
 
       export GOROOT=$(pwd)/go
       export PATH="${GOROOT}/bin:$PATH"
+    else
+      if [ $go_check -eq 2 ] ; then
+        # If go was only just installed previously without shell reset,
+        # makes sure to set correct env variables
+        export GOROOT=/usr/local/go
+        export PATH="${GOROOT}/bin:$PATH"
+      fi
     fi
 
     mkdir gohome
     export GOPATH=$(pwd)/gohome
 
     echo "Checking for git"
+    local git_exists
     if (git version); then
-      local git_exists=1
+      git_exists=1
     else
-      local git_exists=0
+      git_exists=0
 
       sudo apt-get -y -q install git
     fi
@@ -50,25 +68,37 @@ function docker_install_build_relay() {
   cat << 'EOF' >> docker-relay
 #!/bin/bash
 
-connected=$(docker version 2>&1 | grep -c "daemon\|error")
-if [[ ${connected} != 0  ]]; then
+#Import the Windows path
+PATH="$1"
 
-  PATH=${PATH}:$(wslpath "C:\Windows\System32")
-  wHomeWinPath=$(cmd-exe /c 'echo %HOMEDRIVE%%HOMEPATH%' | tr -d '\r')
-  wHome=$(wslpath -u "${wHomeWinPath}")
+# Check if we have Windows Path
+if ( which cmd.exe >/dev/null ); then
 
-  killall --quiet socat
-  exec nohup socat UNIX-LISTEN:/var/run/docker.sock,fork,group=docker,umask=007 EXEC:"\'${wHome}/.npiperelay/npiperelay.exe\' -ep -s //./pipe/docker_engine",nofork  </dev/null &>/dev/null &
+  connected=$(docker version 2>&1 | grep -c "daemon\|error")
+  if [[ ${connected} != 0  ]]; then
+
+    wHomeWinPath=$(cmd-exe /c 'echo %HOMEDRIVE%%HOMEPATH%' | tr -d '\r')
+    wHome=$(wslpath -u "${wHomeWinPath}")
+
+    killall --quiet socat
+    exec nohup socat UNIX-LISTEN:/var/run/docker.sock,fork,group=docker,umask=007 EXEC:"\'${wHome}/.npiperelay/npiperelay.exe\' -ep -s //./pipe/docker_engine",nofork  </dev/null &>/dev/null &
+  fi
 fi
+
 EOF
 
   sudo cp docker-relay /usr/bin/docker-relay
   sudo chmod u+x /usr/bin/docker-relay
 
-  echo '%sudo   ALL=NOPASSWD: /usr/bin/docker-relay' | sudo EDITOR='tee -a' visudo --quiet --file=/etc/sudoers.d/docker-relay
+  echo '%sudo   ALL=NOPASSWD: /usr/bin/docker-relay' | sudo EDITOR='tee ' visudo --quiet --file=/etc/sudoers.d/docker-relay
 
   cat << 'EOF' >> docker_relay.sh
-sudo docker-relay
+
+# Check if we have Windows Path
+if ( which cmd.exe >/dev/null ); then
+  sudo docker-relay "${PATH}"
+fi
+
 EOF
 
   sudo cp docker_relay.sh /etc/profile.d/docker_relay.sh
@@ -98,21 +128,49 @@ EOF
   if [[ ${connected} != 0  ]]; then
     whiptail --title "DOCKER" \
     --msgbox "Please go to Docker Desktop -> Settings -> General and enable 'Expose daemon on tcp://localhost:2375 without TLS' or upgrade your Windows version and run this script again." 9 75
+  else
+    docker version
   fi
 }
 
+function docker_install_conf_toolbox() {
+  echo "Connect to Docker Toolbox"
+
+  cat << 'EOF' >> docker_relay.sh
+
+# Check if we have Windows Path
+if ( which cmd.exe >/dev/null ); then
+  VM=${DOCKER_MACHINE_NAME-default}
+  DOCKER_MACHINE="$(which docker-machine.exe)"
+  eval "$("${DOCKER_MACHINE}" env --shell=bash --no-proxy "${VM}" 2>/dev/null )" > /dev/null 2>&1
+
+  if [[ "${DOCKER_CERT_PATH}" != "" ]] ; then
+    export DOCKER_CERT_PATH="$(wslpath -u "${DOCKER_CERT_PATH}")"
+  fi
+fi
+
+EOF
+  sudo cp docker_relay.sh /etc/profile.d/docker_relay.sh
+
+  . /etc/profile.d/docker_relay.sh
+
+  docker version
+}
+
 function main() {
-  if (whiptail --title "DOCKER" --yesno "Would you like to install the bridge to Docker?" 8 55); then
+  if (confirm --title "DOCKER" --yesno "Would you like to install the bridge to Docker?" 8 55); then
     echo "Installing the bridge to Docker."
 
-    local connected=$(docker.exe version 2>&1 | grep -c "docker daemon is not running.\|docker.exe: command not found")
+    local errorCheck="docker daemon is not running.\|docker.exe: command not found\|error during connect:"
+    local connected
+    connected=$(docker.exe version 2>&1 | grep -c "${errorCheck}")
     while [[ ${connected} != 0  ]]; do
-      if ! (whiptail --title "DOCKER" --yesno "Docker Desktop appears not to be running, please check it and ensure that it is running correctly. Would you like to try again?" 9 75); then
+      if ! (whiptail --title "DOCKER" --yesno "Docker Desktop or Docker Toolbox appears not to be running, please check it and ensure that it is running correctly. Would you like to try again?" 9 75); then
         return
 
       fi
 
-      local connected=$(docker.exe version 2>&1 | grep -c "docker daemon is not running.\|docker.exe: command not found")
+      connected=$(docker.exe version 2>&1 | grep -c "${errorCheck}")
 
     done
 
@@ -120,17 +178,20 @@ function main() {
 
     sudo apt-get -y -q update
 
-    export PATH=${PATH}:$(wslpath "C:\Windows\System32") #Be sure we can execute Windows commands
-
-    wget -c https://download.docker.com/linux/static/stable/$(uname -m)/docker-${DOCKER_VERSION}.tgz
+    wget -c "https://download.docker.com/linux/static/stable/$(uname -m)/docker-${DOCKER_VERSION}.tgz"
     sudo tar -xzvf docker-${DOCKER_VERSION}.tgz --overwrite --directory /usr/bin/ --strip-components 1 docker/docker
 
     sudo chmod 755 /usr/bin/docker
     sudo chown root:root /usr/bin/docker
 
     #Checks if the Windows 10 version supports Unix Sockets and that the tcp port without TLS is not already open
-    local connected=$(env DOCKER_HOST=tcp://0.0.0.0:2375 docker version 2>&1 | grep -c "Cannot connect to the Docker daemon")
-    if [[ $(reg.exe query "HKLM\Software\Microsoft\Windows NT\CurrentVersion" /v "CurrentBuild" 2>&1 | egrep -o '([0-9]{5})' | cut -d ' ' -f 2) -gt 17063 && ${connected} != 0  ]]; then
+    connected=$(env DOCKER_HOST=tcp://0.0.0.0:2375 docker version 2>&1 | grep -c "Cannot connect to the Docker daemon")
+    local currentVersion=$(reg.exe query "HKLM\Software\Microsoft\Windows NT\CurrentVersion" /v "CurrentBuild" 2>&1 | egrep -o '([0-9]{5})' | cut -d ' ' -f 2)
+
+    if [[ $(docker-machine.exe active | grep -c "default") != 0 && ${connected} != 0 ]]; then
+      #Install via Docker Toolbox
+      docker_install_conf_toolbox
+    elif [[ ${currentVersion} -gt 17063 && ${connected} != 0  ]]; then
       #Connect via Unix Sockets
       docker_install_build_relay
     else
@@ -151,7 +212,7 @@ function main() {
 
     docker-compose version
 
-    if [[ $(wslpath 'C:\\') = '/mnt/c/' ]]; then
+    if [[ ${currentVersion} -gt 17063 && $(wslpath 'C:\') = '/mnt/c/' ]]; then
 
       if (whiptail --title "DOCKER" --yesno "To correctly integrate the volume mounting between docker Linux and Windows, your root mount point must be changed from /mnt/c to /c. Continue?" 10 80); then
         echo "Changing the root from /mnt to /"
@@ -171,10 +232,20 @@ for l in $( ls /mnt ); do
     continue
   fi
 
-  if [[ -z $(ls -A /mnt/${l}) ]]; then
+  DEST_PATH=$(wslpath -u "${l^^}:\\" 2>/dev/null)
 
-    rm -d /mnt/${l} #Ensure that we only delete the directory if it is empty
-    ln -s $(wslpath -u "${l^^}:\\") /mnt/${l}
+  if [[ $? != 0 ]]; then
+    continue
+  fi
+
+  if [[ -z $(ls -A /mnt/${l} 2>/dev/null) ]]; then
+
+    if [[ $? != 0 ]]; then
+      continue
+    fi
+
+    rm -d /mnt/${l} 2>/dev/null #Ensure that we only delete the directory if it is empty
+    ln -s $DEST_PATH /mnt/${l} 2>/dev/null
   fi
 
 done
@@ -186,7 +257,12 @@ EOF
         echo '%sudo   ALL=NOPASSWD: /usr/bin/create-mnt-c-link' | sudo EDITOR='tee -a' visudo --quiet --file=/etc/sudoers.d/create-mnt-c-link
 
         cat << 'EOF' >> create-mnt-c-link.sh
-sudo create-mnt-c-link
+
+# Check if we have Windows Path
+if ( which cmd.exe >/dev/null ); then
+  sudo create-mnt-c-link
+fi
+
 EOF
         sudo cp create-mnt-c-link.sh /etc/profile.d/create-mnt-c-link.sh
         sudo chmod -w /usr/bin/create-mnt-c-link
@@ -203,5 +279,5 @@ EOF
   fi
 }
 
-main "$@"
+main
